Crate icu_tutorials

Introduction to ICU4X for Rust

ICU4X is an implementation of Internationalization Components of Unicode (ICU) intended to be modular, performant and flexible.

The library provides a layer of APIs for all software to enable internationalization capabilities.

To use ICU4X in the Rust ecosystem one can either add dependencies on selected components, or add a dependency on the meta-crate icu which brings the full selection of components in the most user-friendly configuration of features.

In this tutorial we are going to build up to writing an app that uses the icu::datetime component to format a date and time, covering various topics in the process.

1. Requirements

For this tutorial we assume the user has basic Rust knowledge. If acquiring it is necessary, the Rust Book provides an excellent introduction. We also assume that the user is familiar with a terminal and have rust and cargo installed.

To verify that, open a terminal and check that the results are similar to:

$ cargo --version
cargo 1.64.0 (387270bc7 2022-09-16)

2. Creating an app with ICU4X as a dependency

Use cargo to initialize a binary application:

cargo new --bin myapp
cd myapp

Then add a dependency on ICU4X’s main crate, icu:

$ cargo add icu

3. Locales

ICU4X comes with a variety of components allowing to manage various facets of software internationalization.

Most of those features depend on the selection of a Locale which is a particular combination of language, script, region with optional variants. An examples of such locales are en-US (American English), sr-Cyrl (Serbian with Cyrillic script) or ar-EG-u-nu-latn (Egyptian Arabic with ASCII numerals).

In ICU4X Locale is a part of the locid component. If the user needs just this one feature, they can use icu_locid crate as a dependency, but since here we already added a dependency on icu, we can refer to it via icu::locid.

Let’s use this in our application.

Open src/main.rs and edit it to:

use icu::locid::Locale;

fn main() {
    let loc: Locale = "ES-AR".parse()
        .expect("Failed to parse locale.");

    if loc.id.language.as_str() == "es" {
        println!("¡Hola!");
    }

    println!("You are using: {}", loc);
}

After saving it, call cargo run and it should display:

¡Hola!
You are using: es-AR

Notice: Here, ICU4X canonicalized the locales’s syntax which uses lowercase letters for the language portion.

Congratulations! ICU4X has been used to semantically operate on a locale!

Convenience macro

The scenario of working with statically declared Locales is common.

It’s a bit unergonomic to have to parse them at runtime and handle a parser error in such case.

For that purpose, ICU4X provides a macro one can use to parse it at compilation time:

use icu::locid::{Locale, locale};

const LOCALE: Locale = locale!("ES-AR");

fn main() {
    if LOCALE.id.language.as_str() == "es" {
        println!("¡Hola amigo!");
    }

    println!("You are using: {}", LOCALE);
}

In this case, the parsing is performed at compilation time, so we don’t need to handle an error case. Try passing an malformed identifier, like “foo-bar” and call cargo check.

Notice: locale! does not support variants or extension tags (e.g. u-nu-latn), as storing these requires allocation. If you have such a tag you need to use runtime parsing.

Next, let’s add some more complex functionality.

4. Basic Data Management

While the locale API is purely algorithmic, many internationalization APIs require more complex data to work. Data management is a complex and non-trivial area which often requires customizations for particular environments and integrations into a project’s ecosystem.

The way ICU4X handles data is one of its novelties, aimed at making the data management more flexible and enabling better integration in asynchronous environments.

As a result, compared to most internationalization solutions, working with data in ICU4X is a bit more explicit. ICU4X provides a trait called DataProvider (as well as BufferProvider and AnyProvider) and a number of concrete APIs that implement these traits for different scenarios. Users are also free to design their own providers that best fit into their ecosystem requirements.

BufferProvider and AnyProvider abstract over different ways the data may be loaded: BufferProvider abstracts over data providers that deserialize data, whereas AnyProvider abstracts over data providers that can provide concrete Rust objects.

5. Using an ICU4X component

We’re going to extend our app to use the icu::datetime component to format a date and time. This component requires data, but as we don’t want to jump into data management just yet, we will use ICU4X’s icu_testdata crate. This contains test providers that support all ICU4X keys for a small representative set of locales. It contains both a BufferProvider (icu_testdata::buffer()), and an AnyProvider (icu_testdata::any()). The latter requires fewer features, so we will be using that.

First, we need to add the crate to our Cargo.toml:

$ cargo add icu_testdata

We can then use it in our code to format a date:

use icu::locid::{Locale, locale};
use icu::calendar::DateTime;
use icu::datetime::{DateTimeFormatter, options::length};

const LOCALE: Locale = locale!("ja"); // let's try some other language

fn main() {
    let provider = icu_testdata::any();

    let options = length::Bag::from_date_time_style(length::Date::Long, length::Time::Medium);

    let dtf = DateTimeFormatter::try_new_with_any_provider(&provider, &LOCALE.into(), options.into())
        .expect("Failed to initialize DateTimeFormatter");

    let date = DateTime::try_new_iso_datetime(2020, 10, 14, 13, 21, 28)
        .expect("Failed to create a datetime.");

    // DateTimeFormatter works with data from any calendar, we need to cast to DateTime<AnyCalendar>
    // For smaller codesize you can use TypedDateTimeFormatter<Gregorian> with a DateTime<Gregorian>
    let date = date.to_any();

    let formatted_date = dtf.format(&date).expect("Formatting should succeed");

    println!("📅: {}", formatted_date);
}

If all went well, running the app with cargo run should display:

📅: 2020年10月14日 13:21:28

Here’s an internationalized date!

Notice: By default, cargo run builds and runs a debug mode of the binary. If you want to evaluate performance, memory or size of this example, use cargo run --release.

6. Summary

This concludes this introduction tutorial. With the help of DateTimeFormat, Locale and DataProvider we formatted a date to Japanese, but that’s just the start. Internationalization is a broad domain and there are many more components in ICU4X.

Next, learn how to generate optimized data for your binary, or continue exploring by reading the docs.

Hooking up a data provider

DataProvider is a general mechanism for loading data required for ICU4X components to operate from a source.

At the moment, DataProvider is only synchronous, but the model of plugging it in is intended to extend to asynchronous DataProviders later.

Data

The first step is to ensure that the provider has a structures to represent the data which will be collected. The structures live in a provider module in your crate and should represent the data efficiently (rather than 1-1 match to CLDR data model).

Types of providers

Any component that needs to use DataProvider should only depend on icu_provider crate and use the DataProvider trait. The specific implementations such as icu_provider_blob::BlobDataProvider will be provided by the downstream consumer of the component.

Hooking up data provider

Each component should use DataProvider only to construct the instance of each main struct that requires data. It means that all heavy data pulling should happen in the constructor, which, in result, must be fallible. Currently, since DataProvider is synchronous, the constructor may be synchronous as well, but in the future we expect to have both synchronous and asynchronous data providers and constructors.

Example

use displaydoc::Display;
use icu_provider::{DataPayload, DataProvider, DataRequest, DataError};
use icu::locid::Locale;
use icu::decimal::provider::{DecimalSymbolsV1Marker, DecimalSymbolsV1};

#[derive(Display, Debug, Copy, Clone)]
pub enum MyError {
     /// Some custom error
     SomeError,

     /// An error originating inside of the data provider.
     #[displaydoc("{0}")]
     DataProvider(DataError),
}

#[cfg(feature = "std")]
impl std::error::Error for MyError {}

impl From<DataError> for MyError {
     fn from(e: DataError) -> Self {
         MyError::DataProvider(e)
     }
}

pub struct AdditiveIdentity(char);

impl AdditiveIdentity {
    pub fn try_new<L: Into<Locale>, P: DataProvider<DecimalSymbolsV1Marker>>(
        locale: L,
        data_provider: &P,
    ) -> Result<Self, MyError> {
        let response = data_provider.load(DataRequest {
            locale: &locale.into().into(),
            metadata: Default::default(),
        })?.take_payload()?;

        let decimal_data: &DecimalSymbolsV1 = response.get();

        Ok(Self(decimal_data.digits[0]))
    }
}

Writing a New Data Struct

ICU4X is a heavily data-driven library. Most new features or components will require pulling in data from an external source.

This tutorial aims to help ICU4X contributors add new data to the data pipeline. It is recommended that readers review data_pipeline.md for additional theory behind the design decisions in the data provider.

Lifecycle of ICU4X Data

It is important to understand the phases of life of ICU4X data as it makes its way from the data source, like CLDR, to the data struct used at runtime. The following flowchart shows the phases and how they connect:

The following steps take place at build time:

Source data file is obtained from an external source. Examples could include the CLDR JSON release or the Unicode Character Database.

The source data is parsed and transformed into a runtime data struct. This step can be expensive, because it is normally run as an offline build step.

The runtime data struct is stored in a way so that a provider can use it: a postcard blob, JSON directory tree, Rust module, etc.

These steps are performed by the icu_datagen, but clients can also write their own data generation logic.

When deserializing from the blob store, it is a design principle of ICU4X that no heap allocations will be required. We have many utilities and abstractions to help make this safe and easy.

Code Layout

With a mental model of the lifecycle of data in ICU4X, we can discuss where to find the code that performs each step.

Data Structs

The data struct definitions should live in the crate that uses them. By convention, the top-level module provider should contain the struct definitions. For example:

icu::decimal::provider::DecimalSymbolsV1

icu::locale_canonicalizer::provider::LikelySubtagsV1

icu::uniset::provider::PropertyCodePointSetV1

In general, data structs should be annotated with #[icu_provider::data_struct], and they should support at least Debug, PartialEq, Clone, Default, and Serde Serialize and Deserialize.

As explained in data_pipeline.md, the data struct should support zero-copy deserialization. The #[icu_provider::data_struct] annotation will enforce this for you. See more information in style_guide.md, as well as the example below in this tutorial.

Data Download

The first step to introduce data into the ICU4X pipeline is to download it from an external source. This corresponds to step 1 above.

When clients use ICU4X, this is generally a manual step, although we may provide tooling to assist with it. For the purpose of ICU4X test data, the tool icu4x-testdata-download-source should automatically download data from the external source and save it in the ICU4X tree. icu4x-testdata-download-source should not do anything other than downloading the raw source data.

To download test data into the ICU4X source tree, run:

$ cargo make testdata-download-sources

Source Data Providers

“Source data providers” read from a source data file, deserialize it, and transform it to an ICU4X data struct. This corresponds to steps 2 and 3 above.

Although they may share common code, source data providers are implemented specific to their data source. There are therefore many source data providers in ICU4X.

Examples of source data providers include:

NumbersProvider

BinaryPropertyCodePointSetDataProvider

… more examples

Source data providers must implement the following traits:

DataProvider<M> or DynamicDataProvider<M> for one or more data markers M; this impl is the main step where data transformation takes place

IterableDataProvider<M>, required for the data exporter (see below)

DynamicDataProvider<SerializeMarker> and IterableDynamicDataProvider<SerializeMarker>, usually implemented with the macro impl_dynamic_data_provider! after the above traits have been implemented

Source data providers are often complex to write. Rules of thumb:

Optimize for readability and maintainability. The source data providers are not used in production, so performance is not a driving concern; however, we want the transformer to be fast enough to make a good developer experience.

If the data source is similar to an existing data source (e.g., importing new data from CLDR JSON), try to share code with existing data providers for that source.

If the data source is novel, feel free to add a new module under icu_datagen::transform.

Data Exporters and Runtime Data Providers

“Data exporters” read from one or more ICU4X data structs and dump them to storage. This corresponds to step 4 above.

Examples of data exporters include:

FilesystemExporter

BlobExporter

“Runtime data providers” are ones that read serialized ICU4X data structs and deserialize them for use at runtime. These are the providers where performance is the key driving factor.

Examples of runtime data providers include:

FsDataProvider

BlobDataProvider

Most ICU4X contributors will not need to touch the data exporters or runtime data providers. New implementations are only necessary when adding a new ICU4X data struct storage mechanism.

Data Generation Tool (icu4x-datagen)

The data generation tool, i.e., icu4x-datagen, ties together the source data providers with a data exporter.

When adding new data structs, it is necessary to make icu4x-datagen aware of your source data provider. To do this, edit provider/datagen/src/registry.rs and add your data provider to the macro

registry!(
    // ...
    FooV1Marker,
)

as well as to the list of keys

use std::borrow::Cow;

#[derive(Debug, PartialEq, Clone)]
#[icu_provider::data_struct(marker(FooV1Marker, "foo/bar@1"))]
pub struct FooV1<'data> {
  message: Cow<'data, str>,
}

When finished, run from the top level:

$ cargo make testdata

If everything is hooked together properly, JSON files for your new data struct should appear under provider/testdata/data/json, and the file provider/testdata/data/testdata.postcard should have changed.

Example

The following example shows all the pieces that make up the data pipeline for DecimalSymbolsV1.

Data Struct

components/decimal/src/provider.rs

use std::borrow::Cow;
use icu_provider::{yoke, zerofrom};
use icu::decimal::provider::{ AffixesV1, GroupingSizesV1 };

/// Symbols and metadata required for formatting a [`FixedDecimal`](crate::FixedDecimal).
#[icu_provider::data_struct(marker(DecimalSymbolsV1Marker, "decimal/symbols@1", extension_key = "nu" ))]
#[derive(Debug, PartialEq, Clone)]
#[cfg_attr(
feature = "datagen",
derive(serde::Serialize, databake::Bake),
databake(path = icu_decimal::provider),
)]
#[cfg_attr(feature = "serde", derive(serde::Deserialize))]
pub struct DecimalSymbolsV1<'data> {
    /// Prefix and suffix to apply when a negative sign is needed.
    #[cfg_attr(feature = "serde", serde(borrow))]
    pub minus_sign_affixes: AffixesV1<'data>,

    /// Prefix and suffix to apply when a plus sign is needed.
    #[cfg_attr(feature = "serde", serde(borrow))]
    pub plus_sign_affixes: AffixesV1<'data>,

    /// Character used to separate the integer and fraction parts of the number.
    #[cfg_attr(feature = "serde", serde(borrow))]
    pub decimal_separator: Cow<'data, str>,

    /// Character used to separate groups in the integer part of the number.
    #[cfg_attr(feature = "serde", serde(borrow))]
    pub grouping_separator: Cow<'data, str>,

    /// Settings used to determine where to place groups in the integer part of the number.
    pub grouping_sizes: GroupingSizesV1,

    /// Digit characters for the current numbering system. In most systems, these digits are
    /// contiguous, but in some systems, such as *hanidec*, they are not contiguous.
    pub digits: [char; 10],
}

The above example is an abridged definition for DecimalSymbolsV1. Note how the lifetime parameter 'data is passed down into all fields that may need to borrow data.

CLDR JSON Deserialize

provider/datagen/src/transform/cldr/cldr_serde/numbers.rs

use icu::locid::LanguageIdentifier;
use itertools::Itertools;
use serde::de::{Deserializer, Error, MapAccess, Unexpected, Visitor};
use serde::Deserialize;
use std::collections::HashMap;
use tinystr::TinyStr8;

#[derive(PartialEq, Debug, Deserialize)]
pub struct Numbers {
    #[serde(rename = "defaultNumberingSystem")]
    pub default_numbering_system: TinyStr8,
    #[serde(rename = "minimumGroupingDigits")]
    #[serde(deserialize_with = "serde_aux::prelude::deserialize_number_from_string")]
    pub minimum_grouping_digits: u8,
}

#[derive(PartialEq, Debug, Deserialize)]
pub struct LangNumbers {
    pub numbers: Numbers,
}

#[derive(PartialEq, Debug, Deserialize)]
pub struct LangData(pub HashMap<LanguageIdentifier, LangNumbers>);

#[derive(PartialEq, Debug, Deserialize)]
pub struct Resource {
    pub main: LangData,
}

The above example is an abridged definition of the Serde structure corresponding to CLDR JSON. Since this Serde definition is not used at runtime, it does not need to be zero-copy.

Transformer

provider/core/src/data_provider.rs

provider/core/src/datagen/iter.rs

impl DataProvider<FooV1Marker> for DatagenProvider {
    fn load(
        &self,
        req: DataRequest,
    ) -> Result<DataResponse<FooV1Marker>, DataError> {
        // Use the data inside self.source and emit it as an ICU4X data struct.
        // This is the core transform operation. This step could take a lot of
        // work, such as pre-parsing patterns, re-organizing the data, etc.
        // This method will be called once per option returned by supported_locales.
    }
}

impl IterableDataProvider<FooV1Marker> for FooProvider {
    fn supported_locales(
        &self,
    ) -> Result<Vec<DataLocale>, DataError> {
        // This should list all supported locales.
    }
}

// Once we have DataProvider and IterableDataProvider, we can apply this macro:
icu_provider::make_exportable_provider!(FooProvider, [FooV1Marker,]);

The above example is an abridged snippet of code illustrating the most important boilerplate for implementing and ICU4X data transform.

Using ICU4X from C++

ICU4X’s core functionality is completely available from C++, with headers generated by Diplomat.

Generated headers can be found under ffi/diplomat/cpp/include, with Sphinx docs at ffi/diplomat/cpp/docs. The port is header-only; no additional C++ translation units need to be compiled to use ICU4X from C++.

Typically C++ users can build ICU4X by building the icu_capi_staticlib crate (crates.io, source) with the cpp_default feature, and link the resultant static library to their C++ application. This crate builds on the icu_capi crate: a no_std crate containing all of the relevant Diplomat-generated extern "C" declarations.

Using ICU4X in C++ is best demonstrated via the examples present in the codebase. For example, here’s an example showing off decimal formatting in ICU4X, built with this Makefile.

We are still working on improving the user experience of using ICU4X from other languages. As such, this tutorial may be a bit sparse, but we are happy to answer questions on our discussions forum and help you out

Building and linking ICU4X

cargo build only supports local crates, so you have to download the published code from crates.io

You can use cargo clone icu_capi_staticlib@1.0.0 to do this (cargo install cargo-clone if the command is not installed)

Inside the crate you can use cargo build to build the library

Be sure to pass --release to get an optimized build

Specify features with --features ...:

buffer_provider for working with blob data providers (ICU4XDataProvider::create_from_byte_slice())

provider_test to include testing data (ICU4XDataProvider::create_test())

provider_fs for loading data from the file system (ICU4XDataProvider::create_fs()). This also requires enabling a syntax on the icu_provider crate.

logging and simple_logger enable basic stdout logging of error metadata. Further loggers can be added on request.

cpp_default turns on a bunch of these and is useful for exploration, but should not be used in production as it enables provider_test

Set CARGO_PROFILE_RELEASE_LTO=true to enable link-time optimization

Set CARGO_PROFILE_RELEASE_OPT_LEVEL="s" to optimize for size

See cargo profiles for more options

Copy the header files from the icu_capi crate

If you’re using the default crate registry, they will be in ~/.cargo/registry/src/*/icu_capi-1.0.0/cpp/include

If you cannot find the crate, download it from crates.io

Make sure to use the same version as above

While we selected a version for icu_capi_staticlib before, and here we’re looking at icu_capi, those versions will always be in sync. You can sanity check this by running cargo pkgid -p icu_capi.

Compile with g++ -std=c++17 icu/target/releaselibicu_capi_staticlib.a -ldl -lpthread -lm. C++ versions beyond C++17 are supported, as are other C++ compilers.

Using ICU4X from C++

Here’s an annotated, shorter version of the fixed decimal example, that can be built using the steps above, using --features cpp_default:

#include "path/to/include/ICU4XFixedDecimalFormatter.hpp"
#include "path/to/include/ICU4XDataStruct.hpp"
#include "path/to/include/ICU4XLogger.hpp"

#include <iostream>
#include <array>

int main() {
   // For basic logging
   ICU4XLogger::init_simple_logger();

   // Create a locale object representing Bangla
   ICU4XLocale locale = ICU4XLocale::create_from_string("bn").ok().value();

   // Use a testing data provider
   // Production code probably should use create_fs() or create_from_byte_slice()
   // with data generated by icu4x-datagen
   ICU4XDataProvider dp = ICU4XDataProvider::create_test();

   // Create a formatter object with the appropriate settings
   ICU4XFixedDecimalFormatter fdf = ICU4XFixedDecimalFormatter::create_with_grouping_strategy(
       dp, locale, ICU4XFixedDecimalGroupingStrategy::Auto).ok().value();

   // Create a decimal representing the number 1,000,007
   ICU4XFixedDecimal decimal = ICU4XFixedDecimal::create_from_u64(1000007);

   // Format it to a string
   std::string out = fdf.format(decimal).ok().value();

   // Report formatted value
   std::cout << "Formatted value is " << out << std::endl;
   if (out != "১০,০০,০০৭") {
       std::cout << "Output does not match expected output" << std::endl;
       return 1;
   }
   return 0;
}

Embedded platforms (no_std)

Users wishing to use ICU4X on a no_std platform will need to write their own crate depending on icu_capi that fills in an allocator and panic hooks, similar to what we do in our example:

#![feature(alloc_error_handler)]

#![no_std]

extern crate icu_capi;
extern crate dlmalloc;

use core::alloc::Layout;
use core::panic::PanicInfo;
use dlmalloc::GlobalDlmalloc;

#[global_allocator]
static ALLOCATOR: GlobalDlmalloc = GlobalDlmalloc;

#[alloc_error_handler]
fn alloc_error(_layout: Layout) -> ! {
    loop {}
}

#[panic_handler]
fn panic(_info: &PanicInfo) -> ! {
    loop {}
}

This can then be compiled with cargo +nightly -Z build-std=std,panic_abort -Z build-std-features=panic_immediate_abort for a minimal build.

Tips

Documentation can be found in the Sphinx docs found ffi/diplomat/cpp/docs. These docs mirror the Rust code in the icu_capi crate, which can be explored on docs.rs, though the precise types used may be different.

Fallible methods return diplomat::result, a Result type that can most commonly be converted to a std::optional over its Ok/Err types by calling .ok() or .err(). Most methods either use ICU4XError (an enum of error codes) as their error type, or std::monostate. Further error details can be logged by enabling a logger via ICU4XLogger, further loggers may be added on request.

The C++ headers include C headers for the underlying APIs as well, under namespace capi, found in the .h files. While these can be used directly, we recommend against it unless you are writing C code. These headers are not intended to be ergonomic and primarily exist for the C++ headers to use internally.

Slices are represented using std::span if available, otherwise a simple wrapper called diplomat::span is used.

These bindings may be customized by running diplomat-tool directly (including replacing the types used with alternate types like mozilla::Span), please ask on our discussions forum for more help on this.

Data management in ICU4X

Unless you’re happy shipping your app with the ~10 locales supported by icu_testdata (and with the size impact of all keys), you will need generate your own data that is customized to suit your needs.

This tutorial introduces data providers beyond icu_testdata, as well as the icu_datagen tool.

1. Prerequesites

This tutorial assumes you have finished the introductory tutorial and continues where that tutorial left off. In particular, you should still have the latest version of code for myapp.

2. Generating data

Data generation is done using the icu_datagen crate, which pulls in data from Unicode’s Common Locale Data Repository (CLDR) and from ICU4C releases to generate ICU4X data. The crate has a command line interface as well as a Rust API, which can be used for example in build.rs files. Here we will use the CLI.

First we will need to install the binary:

$ cargo install icu_datagen --features bin

Get a coffee, this might take a while ☕.

Once installed, run:

$ icu4x-datagen --cldr-tag latest --icuexport-tag latest --out my-data-blob --format blob --all-keys --all-locales

Let’s dissect this invocation:

--cldr-tag selects the CLDR version to use

--icuexport-tag selects the ICU-exported data version to use

--out is the location where we want the generated ICU4X data to be stored

--format sets the format of the output (we’ll discuss formats later)

--all-keys --all-locales specifies that we want to include all data for all locales

This will generate a my-data-blob file containing the serialized data.

3. Using the generated data

Once we have generated some data, it needs to be loaded as a data provider. The blob format we chose can be loaded by BlobDataProvider from the icu_provider_blob crate.

This provider performs deserialization, so it’s a BufferProvider. This means that the feature "serde" needs to be enabled on icu.

Let’s update our Cargo.toml:

$ cargo add icu --features serde
$ cargo add icu_provider_blob

We can then use the provider in our code:

use icu::locid::{locale, Locale};
use icu::calendar::DateTime;
use icu::datetime::{DateTimeFormatter, options::length};
use icu_provider_blob::BlobDataProvider;

const LOCALE: Locale = locale!("ja");

fn main() {
    let blob = std::fs::read("my-data-blob").expect("Failed to read file");
    let buffer_provider = 
        BlobDataProvider::try_new_from_blob(blob.into_boxed_slice())
            .expect("Failed to initialize Data Provider.");

    let options = length::Bag::from_date_time_style(length::Date::Long, length::Time::Medium);

    let dtf = DateTimeFormatter::try_new_with_buffer_provider(&buffer_provider, &LOCALE.into(), options.into())
        .expect("Failed to initialize DateTimeFormatter");

    let date = DateTime::try_new_iso_datetime(2020, 10, 14, 13, 21, 28)
        .expect("Failed to create a datetime.");
    let date = date.to_any();

    let formatted_date = dtf.format(&date).expect("Formatting should succeed");

    println!("📅: {}", formatted_date);
}

Notice: We used the with_buffer_provider constructor (DateTimeFormatter::try_new_with_buffer_provider) now, instead of the with_any_provider constructor that we used with test data, as we now have a BufferProvider.

4. Data slicing

You might have noticed that the blob we generated is a hefty 13MB. This is no surprise, as we included --all-keys --all-locales. However, our binary only uses date formatting data in Japanese. There’s room for optimization:

$ icu4x-datagen --overwrite --cldr-tag latest --icuexport-tag latest --out my-data-blob --format blob --keys-for-bin target/debug/myapp --locales ja

The --keys-for-bin argument tells icu4x-datagen to analyze the binary and only include keys that are used by its code. In addition, we know that we only need data for the Japanese locale. This significantly reduces the blob’s file size, to 54KB, and our program still works. Quite the improvement!

But there is more to optimize. You might have noticed this in the output of the icu4x-datagen invocation, which lists 21 keys, including clearly irrelevant ones like datetime/ethopic/datesymbols@1. Remember how we had to convert our DateTime<Gregorian> into a DateTime<AnyCalendar> in order to use the DateTimeFormatter? Turns out, as DateTimeFormatter contains logic for many different calendars, datagen includes data for all of these as well.

We can instead use TypedDateTimeFormatter<Gregorian>, which only supports formatting DateTime<Gregorian>s:

use icu::locid::{locale, Locale};
use icu::calendar::{DateTime, Gregorian};
use icu::datetime::{TypedDateTimeFormatter, options::length};
use icu_provider_blob::BlobDataProvider;

const LOCALE: Locale = locale!("ja");

fn main() {
    let blob = std::fs::read("my-data-blob").expect("Failed to read file");
    let buffer_provider = 
        BlobDataProvider::try_new_from_blob(blob.into_boxed_slice())
            .expect("Failed to initialize Data Provider.");

    let options = length::Bag::from_date_time_style(length::Date::Long, length::Time::Medium);

    let dtf = TypedDateTimeFormatter::<Gregorian>::try_new_with_buffer_provider(&buffer_provider, &LOCALE.into(), options.into())
        .expect("Failed to initialize DateTimeFormatter");

    let date = DateTime::try_new_gregorian_datetime(2020, 10, 14, 13, 21, 28)
        .expect("Failed to create a datetime.");

    let formatted_date = dtf.format(&date);

    println!("📅: {}", formatted_date);
}

This has two advantages: it reduces our code size, as DateTimeFormatter includes much more functionality than TypedDateTimeFormatter<Gregorian>, and it reduces our data size, as static analysis can now determine that we need even fewer keys. The data size improvement could have also been achieved by manually listing the data keys we think we’ll need (there’s a --keys flag), but we risk a runtime error if we’re wrong.

This is a common pattern in ICU4X, and most of our APIs are designed with data slicing in mind.

Rebuilding the application and rerunning datagen awards us with a 3KB data blob, which only contains 7 data keys!

5. Other formats

So far we’ve used --format blob and BlobDataProvider. This is useful if we want to ship code and data separately, but there are other options.

mod and baked data

The mod format will generate a Rust module that contains all the required data directly as Rust code. This format naturally has no deserialization overhead, and allows for compile-time optimizations (data slicing isn’t really necessary, as the compiler will do it for us), but cannot be dynamically loaded at runtime.

Let’s give it a try:

$ icu4x-datagen --cldr-tag latest --icuexport-tag latest --out my-data-mod --format mod --keys-for-bin target/debug/myapp --locales ja

The output might tell you additional crates that need to be installed. Don’t worry, these are transitive dependencies already anyway, but are required directly now to construct our data:

$ cargo add icu_provider
$ cargo add litemap
$ cargo add zerovec

We can then use the data by directly including the source with the include! macro.

extern crate alloc; // required as my-data-mod is written for #[no_std]
use icu::locid::{locale, Locale};
use icu::calendar::DateTime;
use icu::datetime::{TypedDateTimeFormatter, options::length};

const LOCALE: Locale = locale!("ja");

struct UnstableProvider;
include!("../my-data-mod/mod.rs");
impl_data_provider!(UnstableProvider);

fn main() {
    let unstable_provider = UnstableProvider;

    let options = length::Bag::from_date_time_style(length::Date::Long, length::Time::Medium);

    let dtf = TypedDateTimeFormatter::try_new_unstable(&unstable_provider, &LOCALE.into(), options.into())
        .expect("Failed to initialize DateTimeFormatter");

    let date = DateTime::try_new_gregorian_datetime(2020, 10, 14, 13, 21, 28)
        .expect("Failed to create a datetime.");

    let formatted_date = dtf.format(&date);

    println!("📅: {}", formatted_date);
}

With this provider, we can use the unstable constructors. These are only guaranteed to work if the data was generated with the same version of ICU4X that you are building with, but if you build the data as part of your a build pipeline, that shouldn’t be a problem.

You can also implement the AnyProvider trait, so that it can be used with _with_any_provider constructors. Using these constructors is slightly less performant than the unstable ones, but, as the name suggests, stable across (minor) releases.

impl_any_provider!(MyProvider);

dir and FsDataProvider

The dir format will generate a directory tree of data files in JSON (although the --syntax option can be used to generate postcard or bincode files, which doesn’t have many practical uses).

Let’s give it a try:

$ icu4x-datagen --cldr-tag latest --icuexport-tag latest --out my-data-dir --format dir --keys-for-bin target/debug/myapp --locales ja

This directory can be read by the FsDataProvider from the icu_provider_fs crate. You will also need to activate the feature for the chosen syntax on the icu_provider crate.

Same as BlobDataProvider, this also a buffer provider, so you will need to activate icu’s serde feature and use the with_buffer_provider constructors.

$ cargo add icu --features serde
$ cargo add icu_provider --features deserialize_json
$ cargo add icu_provider_fs

use icu::locid::{locale, Locale};
use icu::calendar::DateTime;
use icu::datetime::{TypedDateTimeFormatter, options::length};
use icu_provider_fs::FsDataProvider;

const LOCALE: Locale = locale!("ja");

fn main() {
    let buffer_provider = FsDataProvider::try_new("my-data-dir")
           .expect("Failed to initialize Data Provider");

    let options = length::Bag::from_date_time_style(length::Date::Long, length::Time::Medium);

    let dtf = TypedDateTimeFormatter::try_new_with_buffer_provider(&buffer_provider, &LOCALE.into(), options.into())
        .expect("Failed to initialize DateTimeFormatter");

    let date = DateTime::try_new_gregorian_datetime(2020, 10, 14, 13, 21, 28)
        .expect("Failed to create a datetime.");

    let formatted_date = dtf.format(&date);

    println!("📅: {}", formatted_date);
}

6. Summary

We have learned how to generate data and load it into our programs, optimize data size, and gotten to know the different data providers that are part of ICU4X.

You can learn more about datagen, including the Rust API which we have not used in this tutorial, by reading the docs.

Using ICU4X from JavaScript and TypeScript

ICU4X’s core functionality is completely available from JS and TypeScript via WebAssembly, with bindings generated by Diplomat.

A sample NPM package can be found under ffi/diplomat/js/examples/node, with Sphinx docs at ffi/diplomat/js/docs. You can also run tsdoc on the package to get TypeScript docs.

We are still working on improving the user experience of using ICU4X from other languages. As such, this tutorial may be a bit sparse, but we are happy to answer questions on our discussions forum and help you out

Building the package

The example package contains test data, which is why it’s not released on NPM. You will need to build the package yourself to select the desired Rust features.

See [ffi/diplomat/js/examples/node/package.json] for inspiration.

Using the built package

Similar to C++, the JS APIs mirror the Rust code in the icu_capi crate, which can be explored on docs.rs, though the precise types used may be different.

See ffi/diplomat/js/examples/wasm-demo for an NPM package that uses the example ICU4X package. You can run it using npm run start.

We hope to fill in these docs over time with more examples.

Techniques for Writing Design Docs

According to bylaws.md, designs should make their way to GitHub Markdown files somewhat early in the process. However, since Markdown is not a WYSIWYG platform like Google Docs, it takes a bit of time before you become accustomed to the practice. This document contains tips on the workflow.

Design Document Evolution

Google Doc for Rapid Prototyping

A first step when writing a design doc may be to start in Google Docs. You can share the doc with a trusted group to receive early feedback on typographical errors, mistakes, points that need clarification, and so on. You should also be keeping an eye out for open questions, and you should explicitly enumerate open questions in your doc.

This phase is called “Rapid Prototyping” as defined by the Bylaws.

Markdown Draft

Once you’ve done some basic polishing of your design doc, you should migrate it to Markdown in the ICU4X repository. You should file issues for all open questions, and reference those open questions from your doc. Send the Markdown file out for review in a PR, and we will discuss it in the next ICU4X-SC weekly meeting. If the SC agrees with the design you have put forward in the doc, the PR will be considered as approved and can be merged in.

Addressing Open Questions

Once the working draft of your design is checked in, you can address open questions using the GitHub issue tracking feature. Once each question reaches a resolution, open a PR referencing that issue, updating the Markdown file to reflect the conclusion. Once approved by the SC, the PR can be merged and the issue can be closed.

Markdown Techniques

Images

You can include images using code such as the following:

![JSON Data Provider](../assets/multi_provider_architecture.svg)

Inside the ![...] is the alt text, and inside the (...) is the relative path to the image file. You should save your images in an assets folder in the same directory as the Markdown file.

Shane likes SVG files because they scale up and down nicely on different screen sizes and resolutions. However, bitmaps are acceptable, too, depending on how you created the images.

Create the images using your favorite editor. For the images in data-pipeline.md, Shane used Lucidchart. You can also use Inkscape, Gimp, Google Slides, Adobe Illustrator, etc. If you have a tablet, another good option may be to draw the image on your screen with your stylus into a whiteboard program and export it as an image.

Previewing Images in PRs

Here is how others can view your Markdown with images in a PR:

In the PR, click “Files Changed” (goes to /pull/123/files)

Find the Markdown file you want to preview

Click the three dots (…) to open the menu, then click “View File” (goes to /blob/abcdef0/docs/example.md)